pyproject.toml
[project]
name = "learning-alert-triage"
version = "0.1.0"
description = "Research project for AI-powered security alert triage continuous learning"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "langfuse==3.2.4",
    "requests>=2.32.5",
    "google-genai>=0.1.0",
    "streamlit>=1.31.0",
    "pandas>=2.2.0",
    "plotly>=5.18.0",
]

[tool.uv]


